Refactor

+++ src/components/ViewRestaurantModalContent.tsx
@@ -1,6 +1,8 @@
 import { deleteDoc, doc } from "firebase/firestore";
 import { db } from "../services/firebase";
 import { IRestaurant } from "../services/types";
+import { ThumbsDownIcon, ThumbsUpIcon } from "./Icons";
+import StarRating from "./StarRating";
 
 type ViewProps = {
   restaurant: IRestaurant;
@@ -14,43 +16,61 @@ export default function ViewRestaurantModalContent({
   setIsView,
 }: ViewProps) {
   const deleteRestaurant = async () => {
-    deleteDoc(doc(db, "restaurants", restaurant.id));
+    if (window.confirm("Are you sure you want to delete this restaurant?")) {
+      deleteDoc(doc(db, "restaurants", restaurant.id));
+    }
   };
 
   return (
     <div className="modal-box h-3/4">
-      <div className="flex flex-row mb-4">
-        <h1 className="text-2xl mr-4">{restaurant.name}</h1>
-        <button className="btn btn-sm" onClick={() => setIsView(!isView)}>
-          Edit
-        </button>
+      <div className="flex flex-row mb-8 justify-between">
+        <div className="flex flex-row">
+          <h1 className="text-2xl mr-2">{restaurant.name}</h1>
+          <button
+            className="btn btn-sm mr-4"
+            onClick={() => setIsView(!isView)}
+          >
+            Edit
+          </button>
+        </div>
+        <form method="dialog">
+          <button
+            onClick={deleteRestaurant}
+            className="btn btn-sm btn-outline btn-error"
+            type="submit"
+          >
+            Delete
+          </button>
+        </form>
       </div>
       <div className="flex flex-col mb-4">
         <div className="flex flex-row justify-between">
-          <p>Food Rating: {restaurant.foodRating}</p>
-          <p>Service Rating: {restaurant.serviceRating}</p>
+          <div>
+            <p>Food</p>
+            <StarRating rating={restaurant.foodRating} />
+          </div>
+          <div>
+            <p>Service</p>
+            <StarRating rating={restaurant.serviceRating} />
+          </div>
         </div>
         <div>
-          <p>Vibe Rating: {restaurant.vibeRating}</p>
+          <p>Vibe</p>
+          <StarRating rating={restaurant.vibeRating} />
         </div>
       </div>
       <div>
+        <p className="mb-2">Menu Items</p>
         {restaurant.menuItems.map((mI) => (
-          <div key={mI.id} className="flex flex-row items-center">
-            <p>{mI.name}</p>
-            <p>{mI.wouldEatAgain ? "Yes" : "No"}</p>
+          <div
+            key={mI.id}
+            className="flex flex-row items-center p-2 bg-neutral mb-3 rounded-lg w-fit"
+          >
+            <p className="mr-4">{mI.name}</p>
+            {mI.wouldEatAgain ? <ThumbsUpIcon /> : <ThumbsDownIcon />}
           </div>
         ))}
       </div>
-      <form method="dialog">
-        <button
-          onClick={deleteRestaurant}
-          className="btn btn-sm btn-outline btn-error mt-8"
-          type="submit"
-        >
-          Delete Restaurant
-        </button>
-      </form>
     </div>
   );
 }
